+++ src/test/pages/QuizAttempt.test.tsx
@@ -1,20 +1,14 @@
-import {act, screen, within} from "@testing-library/react";
-import { expectLinkWithEnabledBackwardsNavigation, expectH1, expectH4, expectTextInElementWithId, expectTitledSection, expectUrl, renderTestEnvironment, setUrl, waitForLoaded } from "../testUtils";
+import { expectLinkWithEnabledBackwardsNavigation, expectH1, expectH4, expectTitledSection, expectUrl } from "../testUtils";
 import {mockAttempts} from "../../mocks/data";
-import type {UserRole} from "../../IsaacApiTypes";
 import { siteSpecific } from "../../app/services";
+import { expectActionMessage, expectBreadcrumbs, expectErrorMessage, renderQuizPage, testSectionsHeader } from "../helpers/quiz";
 
 describe("QuizAttempt", () => {
     const quizId = Object.keys(mockAttempts)[0];
     const attempt = mockAttempts[quizId];
     const sections = attempt.quiz?.children;
 
-    const renderQuizAttempt =  async ({role, quizId}: {role: UserRole | "ANONYMOUS", quizId: string}) => {
-        await act(async () => renderTestEnvironment({ role }));
-        await act(async () => setUrl({ pathname: `/test/attempt/${quizId}` }));
-        await waitForLoaded();
-    };
-
+    const renderQuizAttempt = renderQuizPage('/test/attempt');
     const studentAttemptsQuiz = () => renderQuizAttempt({ role: 'STUDENT', quizId });
 
     describe("overview", () => {
@@ -97,15 +91,3 @@ describe("QuizAttempt", () => {
         });
     });
 });
-
-const expectErrorMessage = expectTextInElementWithId('error-message');
-const expectActionMessage = expectTextInElementWithId('quiz-action');
-const testSectionsHeader = () => screen.queryByRole('heading', {name: "Test sections"});
-const expectBreadcrumbs = ([first, second, third]: [{href: string, text: string}, {href: string, text: string}, string | undefined]) => {
-    const breadcrumbs = within(screen.getByRole('navigation', { name: 'breadcrumb' })).getByRole('list');
-    expect(Array.from(breadcrumbs.children).map(e => e.innerHTML)).toEqual([
-        `<a href="${first.href}"><span>${first.text}</span></a>`,
-        `<a href="${second.href}"><span>${second.text}</span></a>`,
-        `<span>${third}</span>`
-    ]);
-};
